fix: play moooler sound when CoW order is submitted, not on review open
